+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- (Safari) Fix lookup puck taps and double-taps on recent iOS 26 releases.
+
 ## [1.0.1] - 2024-11-20
 
 ### Changed